+++ recipes/client.rb
@@ -36,8 +36,6 @@ else
   masters.each do |master|
     node.default['chrony']['servers'][master['ipaddress']] = master['chrony']['server_options']
     node.default['chrony']['allow'].push "allow #{master['ipaddress']}"
-    # only use 1 server to sync initslewstep
-    node.default['chrony']['initslewstep'] = "initslewstep 20 #{master['ipaddress']}"
   end
 end