+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of the chrony cookbook.
 - Sous Chefs Adoption
 - Standardize helpers library
 - Improve testing
+- Remove unused attributes
 
 ## 0.3.1 (2020-12-23)